+++ jest.config.js
@@ -51,7 +51,8 @@ module.exports = {
         testsForPackage('plugin-react-native-client-sync'),
         testsForPackage('plugin-react-native-global-error-handler'),
         testsForPackage('plugin-react-native-event-sync'),
-        testsForPackage('plugin-react-native-session')
+        testsForPackage('plugin-react-native-session'),
+        testsForPackage('plugin-react-native-navigation')
       ],
       setupFiles: [
         require.resolve('react-native/Libraries/Core/setUpGlobals.js'),